+++ app/Http/Resources/ArtistSearchResultResource.php
@@ -24,7 +24,9 @@ class ArtistSearchResultResource extends JsonResource
             'name' => $this->name,
             'genres' => $this->genres,
             'image_url' => $this->imageUrl,
+            'spotify_popularity' => $this->popularity,
             'exists_in_database' => $this->existsInDatabase,
+            'score' => $this->score,
             'source' => $this->source,
         ];
     }